requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pandas
openpyxl
langchain
langchain-groq
langchain-openai
